Add intermediate build recipes and improve multi target support
dk-generate now take multiple targets instead of just one,
and will output the build recipes for all of them.
A .build.yaml file describes the different build recipes,
the files they consists of, which toolchain they are designed
for, etc.
This file is passed to dk-build, allowing it to have all the
information needed to process the build recipes correctly.